Revert "Factor test runner into `run_mech_tests` with CLI `--out` support and JSON/MEC output"
